switch from using int to long in IDataProvider and the cascade from that Issue: WICKET-1175
git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@1203602 13f79535-47bb-0310-9956-ffa450edef68

+++ wicket-core/src/main/java/org/apache/wicket/markup/repeater/data/DataView.java
@@ -89,7 +89,7 @@ public abstract class DataView<T> extends DataViewBase<T>
 	 * @param itemsPerPage
 	 *            items per page
 	 */
-	protected DataView(String id, IDataProvider<T> dataProvider, int itemsPerPage)
+	protected DataView(String id, IDataProvider<T> dataProvider, long itemsPerPage)
 	{
 		super(id, dataProvider);
 		setItemsPerPage(itemsPerPage);
